+++ handler/api/api.go
@@ -28,6 +28,7 @@ import (
 	"github.com/drone/drone/handler/api/repos"
 	"github.com/drone/drone/handler/api/repos/builds"
 	"github.com/drone/drone/handler/api/repos/builds/branches"
+	"github.com/drone/drone/handler/api/repos/builds/deploys"
 	"github.com/drone/drone/handler/api/repos/builds/logs"
 	"github.com/drone/drone/handler/api/repos/builds/pulls"
 	"github.com/drone/drone/handler/api/repos/builds/stages"
@@ -191,6 +192,9 @@ func (s Server) Handler() http.Handler {
 
 				r.Get("/pulls", pulls.HandleList(s.Repos, s.Builds))
 				r.With(acl.CheckWriteAccess()).Delete("/pulls/{pull}", pulls.HandleDelete(s.Repos, s.Builds))
+
+				r.Get("/deployments", deploys.HandleList(s.Repos, s.Builds))
+				r.With(acl.CheckWriteAccess()).Delete("/deployments/*", deploys.HandleDelete(s.Repos, s.Builds))
 
 				r.Get("/latest", builds.HandleLast(s.Repos, s.Builds, s.Stages))
 				r.Get("/{number}", builds.HandleFind(s.Repos, s.Builds, s.Stages))